requirements.txt
django
pyrogram
bs4
lxml
django-ckeditor
tgcrypto
whitenoise